refactor: :recycle: split add modal
